Add files via upload
Adding .rmd and .pdf of the completed and 'ran' .rmd file, showing my output with the data. Without the dependent .csv's, the code cannot be executed, however it helps transparency to see the code used, and the pdf substantiates this. If cleared by releveant supervisors, the data is anonymous so will hopefully be uploaded to follow, where readers/researchers may run their own analyses if they see fit.